+++ core/src/test/scala/unit/kafka/admin/ResetConsumerGroupOffsetTest.scala
@@ -232,7 +232,7 @@ class ResetConsumerGroupOffsetTest extends ConsumerGroupCommandTest {
   @Test
   def testResetOffsetsToEarliestOnOneTopicAndPartition() {
     val topic = "bar"
-    adminZkClient.createTopic(topic, 2, 1)
+    createTopic(topic, 2, 1)
 
     val args = Array("--bootstrap-server", brokerList, "--reset-offsets", "--group", group, "--topic",
       s"$topic:1", "--to-earliest", "--execute")
@@ -253,8 +253,8 @@ class ResetConsumerGroupOffsetTest extends ConsumerGroupCommandTest {
   def testResetOffsetsToEarliestOnTopics() {
     val topic1 = "topic1"
     val topic2 = "topic2"
-    adminZkClient.createTopic(topic1, 1, 1)
-    adminZkClient.createTopic(topic2, 1, 1)
+    createTopic(topic1, 1, 1)
+    createTopic(topic2, 1, 1)
 
     val args = Array("--bootstrap-server", brokerList, "--reset-offsets", "--group", group, "--topic", topic1,
       "--topic", topic2, "--to-earliest", "--execute")
@@ -280,8 +280,8 @@ class ResetConsumerGroupOffsetTest extends ConsumerGroupCommandTest {
     val topic1 = "topic1"
     val topic2 = "topic2"
 
-    adminZkClient.createTopic(topic1, 2, 1)
-    adminZkClient.createTopic(topic2, 2, 1)
+    createTopic(topic1, 2, 1)
+    createTopic(topic2, 2, 1)
 
     val args = Array("--bootstrap-server", brokerList, "--reset-offsets", "--group", group, "--topic",
       s"$topic1:1", "--topic", s"$topic2:1", "--to-earliest", "--execute")
@@ -310,7 +310,7 @@ class ResetConsumerGroupOffsetTest extends ConsumerGroupCommandTest {
     val topic = "bar"
     val tp0 = new TopicPartition(topic, 0)
     val tp1 = new TopicPartition(topic, 1)
-    adminZkClient.createTopic(topic, 2, 1)
+    createTopic(topic, 2, 1)
 
     val cgcArgs = Array("--bootstrap-server", brokerList, "--reset-offsets", "--group", group, "--all-topics",
       "--to-offset", "2", "--export")
